Rename pubserver module directory.
