+++ src/pubd/manager.rs
