+++ src/pubd/mod.rs
@@ -1,7 +1,7 @@
 mod commands;
 mod events;
+mod manager;
 mod publishers;
-mod pubserver;
 mod repository;
 
 pub use self::commands::{RepoAccessCmd, RepoAccessCmdDet};
@@ -9,6 +9,6 @@ pub use self::events::{
     RepositoryAccessEvent, RepositoryAccessEventDetails, RepositoryAccessIni, RepositoryAccessInitDetails,
     RrdpSessionReset, RrdpUpdate,
 };
+pub use self::manager::RepositoryManager;
 pub use self::publishers::Publisher;
-pub use self::pubserver::RepositoryManager;
 pub use self::repository::*;
